+++ domi1819.UpClient/Forms/InfoForm.cs
@@ -24,7 +24,7 @@ namespace domi1819.UpClient.Forms
             get
             {
                 CreateParams createParams = base.CreateParams;
-                createParams.ExStyle |= (int)WinConsts.WS_EX_TOPMOST;
+                createParams.ExStyle |= WinConsts.WS_EX_TOPMOST;
 
                 return createParams;
             }
@@ -48,14 +48,14 @@ namespace domi1819.UpClient.Forms
 
         protected override void WndProc(ref Message m)
         {
-            switch ((uint)m.Msg)
+            switch (m.Msg)
             {
                 case WinConsts.WM_NCHITTEST:
-                    m.Result = new IntPtr((int)WinConsts.HTCLIENT);
+                    m.Result = new IntPtr(WinConsts.HTCLIENT);
                     return;
 
                 case WinConsts.WM_NCACTIVATE:
-                    m.WParam = new IntPtr((int)WinConsts.TRUE);
+                    m.WParam = new IntPtr(WinConsts.TRUE);
                     break;
             }
 
@@ -66,7 +66,7 @@ namespace domi1819.UpClient.Forms
         {
             if (!this.showing)
             {
-                Message m = new Message { HWnd = this.Handle, Msg = (int)WinConsts.WM_NCACTIVATE };
+                Message m = new Message { HWnd = this.Handle, Msg = WinConsts.WM_NCACTIVATE };
                 
                 this.WndProc(ref m);
